+++ .env
@@ -1,5 +1,5 @@
 GOOGLE_CLIENT_ID="724966659273-5gji8f5cgei0v7oudkoj44j5mt6r8ibq.apps.googleusercontent.com"
 GOOGLE_CLIENT_SECRET="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"
 
-GOOGLE_API_KEY=AIzaSyB98vvtxKPgOT-0YhKtN_CdJVY3F3KFqkI
+REACT_APP_GOOGLE_API_KEY=AIzaSyB98vvtxKPgOT-0YhKtN_CdJVY3F3KFqkI
 NEXT_PUBLIC_GOOGLE_API_KEY=AIzaSyB98vvtxKPgOT-0YhKtN_CdJVY3F3KFqkI
