+++ src/app/components/constants.js
@@ -1,0 +1,1 @@
+REACT_APP_GOOGLE_API_KEY="AIzaSyB98vvtxKPgOT-0YhKtN_CdJVY3F3KFqkI"
